Allow exporting SVG of buggy boards
Some boards, like ones based on the "Quick retrospective" template
seem to have a race condition in Miro itself. The entity management system
in Miro tracks "broken entities", which these boards trigger for the type
32, but in certain cases the Miro socket sends the 6loaded_FAL command
after this broken entity has been registered. This command clears the list
of broken entities, and as such the board never finishes loading because
this permanently broken entity will never load and the entity manager has
lost the information about the broken entity. It is unknown how this should
be properly addressed by Miro, but either the clearing of the broken
entities mechanism is faulty (it should not clear them after receiving
`6loaded_FAL`) or the broken 32 entity should be not broken.  In any case, the
underlying APIs needed to export the SVG are available even though the board UI
hasn't finished loading. However, the other APIs (i.e., finding objects by
filters and selecting certain objects) do need the board to fully load. As
such, exporting the board to JSON or exporting certain objects of the board
won't work; only exporting the entire board as a SVG works using this
workaround.

Related: #65